html: UnescapeString: fix an off-by-1 error
There is an off-by-1 error where it checks "<= 3" instead of "< 3", so
it doesn't accept numeric entities that are 3 characters; which means
single-digit decimal-encoded values that don't have a semicolon.

This erroneous check occurs both pre-flight and post-flight.  The fix
for the pre-flight check is simple: just change it to "< 3".  The
post-flight check needs to be taught to differentiate between decimal
and hexadecimal values.