html: UnescapeString: detect integer overflow
When reading numeric entities, UnescapeString reads the number into a
rune (32-bit integer), but does not detect 32-bit overflow.  This
allows an entity such as "&#x10000000000000061;" to be interpreted as
the ASCII character "a", rather than noting that it's above
unicode.RunMax (U+10FFFF) and interpreting it as
unicode.ReplacementChar (U+FFFD).

To solve this: First, simply make a note of it when we cross U+10FFFF.
Then after we're done scanning in the number, if we made that note,
then set the value to a fixed above-MaxRune value.